tools/ocr_tts/requirements.txt
Pillow
pytesseract
gTTS
